+++ packages/profile-sync-controller/src/controllers/user-storage/UserStorageController.ts
@@ -2,6 +2,7 @@ import type {
   AccountsControllerListAccountsAction,
   AccountsControllerUpdateAccountMetadataAction,
   AccountsControllerAccountRenamedEvent,
+  AccountsControllerAccountAddedEvent,
   AccountsControllerUpdateAccountsAction,
 } from '@metamask/accounts-controller';
 import type {
@@ -45,6 +46,7 @@ import {
   type UserStoragePathWithFeatureOnly,
 } from '../../shared/storage-schema';
 import type { NativeScrypt } from '../../shared/types/encryption';
+import { EventQueue } from '../../shared/utils/event-queue';
 import { createSnapSignMessageRequest } from '../authentication/auth-snap-requests';
 import type {
   AuthenticationControllerGetBearerToken,
@@ -256,6 +258,7 @@ export type AllowedEvents =
   | KeyringControllerUnlockEvent
   // Account Syncing Events
   | AccountsControllerAccountRenamedEvent
+  | AccountsControllerAccountAddedEvent
   // Network Syncing Events
   | NetworkControllerNetworkRemovedEvent;
 
@@ -331,6 +334,8 @@ export default class UserStorageController extends BaseController<
   };
 
   readonly #nativeScryptCrypto: NativeScrypt | undefined = undefined;
+
+  eventQueue = new EventQueue();
 
   constructor({
     messenger,